requirements.txt
python-telegram-bot[webhooks]==20.0
aiohttp==3.9.0
aiosqlite==0.19.0
beautifulsoup4==4.12.2
lxml==4.9.4
cachetools==5.3.2
uvloop==0.19.0
typing-extensions==4.8.0
